Refactored the category spending files
