+++ src/finance_modules/SpendingCategoryCharts.jsx
